pyproject.toml
[tool.poetry]
name = "good-food"
version = "0.1.0"
description = "Healthy food online store"
authors = ["Healthy Food <healthyfoodapi@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.2.0"}
pillow = "^10.1.0"
django-cleanup = "^8.0.0"
djoser = "^2.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pep8-naming = "^0.13.3"
flake8-broken-line = "^1.0.0"
flake8-return = "^1.2.0"
flake8-isort = "^6.1.0"
isort = "^5.12.0"
pytest = "^7.4.2"
black = "^23.10.0"
django-debug-toolbar = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
